+++ src/adapter/index.ts
@@ -6,8 +6,7 @@ const {WORK_PATH} = botInfo
 
 export async function Plugin() {
     try {
-        const { Plugin } = await import('yunzai')
-        return Plugin
+        return global.plugin
     } catch (err) {
         try {
             const plugin = (await import(join(WORK_PATH,'lib/plugins/plugin.js'))).default
@@ -21,9 +20,6 @@ export async function Plugin() {
 
 export async function Puppeteer() {
     try {
-        const { puppeteer } = await import('yunzai')
-        return puppeteer
-    } catch (err) {
         try {
             const puppeteer = (await import(join(WORK_PATH,'lib/puppeteer/puppeteer.js'))).default
             return puppeteer
@@ -31,24 +27,24 @@ export async function Puppeteer() {
             const puppeteer = (await import(pathToFileURL(join(WORK_PATH,'lib/puppeteer/puppeteer.js')).toString())).default
             return puppeteer
         }
+    } catch (err) {
+        const { puppeteer } = await import('yunzaijs')
+        return puppeteer
     }
 }
 
 export async function Segment():Promise<typeof global.segment> {
     try {
-        const Segment = global.segment
-        return Segment
+        return global.segment
     } catch (err) {
-        const { Segment } = await import('yunzai')
+        const { Segment } = await import('yunzaijs')
         return Segment
     }
 }
 
 export async function Loader() {
     try {
-        const { Loader } = await import("yunzai")
-        return Loader
-    } catch (err) {
+        
         try {
             const loader = (await import(join(WORK_PATH,"lib/plugins/loader.js"))).default
             return loader
@@ -56,34 +52,20 @@ export async function Loader() {
             const loader = (await import(pathToFileURL(join(WORK_PATH,"lib/plugins/loader.js")).toString())).default
             return loader
         }
+    } catch (err) {
+        const { Loader } = await import("yunzaijs")
+        return Loader
     }
 }
 
 export async function Bot():Promise<typeof global.Bot> {
-    try {
-        const bot = global.Bot
-        return bot
-    } catch (err) {
-        const { Bot } = await import('yunzai')
-        return Bot as typeof global.Bot
-    }
+    return global.Bot
 }
 
 export async function Logger():Promise<typeof global.logger> {
-    try {
-        const logger = global.logger
-        return logger
-    } catch (err) {
-
-    }
+    return global.logger
 }
 
 export async function Redis():Promise<import("redis").RedisClientType> {
-    try {
-        const redis = global.redis
-        return redis
-    } catch (err) {
-        const { Redis } = await import('yunzai')
-        return Redis as import("redis").RedisClientType
-    }
+    return global.redis
 }