+++ src/main.ts
@@ -1,4 +1,4 @@
-import { Client, createLogin, Processor } from 'yunzai'
+import { Client, createLogin, Processor } from 'yunzaijs'
 setTimeout(async () => {
   // 输入login配置
   await createLogin()