+++ src/server/app/webui/terminal/index.ts
@@ -1,7 +1,8 @@
 import iconvLite from 'iconv-lite'
+import os from 'os'
 import { randomUUID } from "crypto"
 import { WebSocket } from 'ws'
-import { exec } from "child_process"
+import { spawn } from "child_process"
 import { resolve as resolvePath } from 'path'
 import { Logger } from '#bot'
 import { TermCfg } from './config'
@@ -58,26 +59,67 @@ class TerminalWs {
             ws.send(JSON.stringify({ type: 'exec', action: 'stdout', params: data }))
         }
 
-        if (cmd.trim() !== 'exit') {
+        if (cmd.trim() !== 'exit') { 
 
-            exec(cmd, { 
-                encoding: 'buffer',
-                cwd: this.execPath
-            }, (error, stdout, stderr) => {  
-                if (error) {  
-                    // logger.error(`exec error: ${iconvLite.decode(Buffer.from(error.message), 'cp936')}`);  
-                    logger.error(error.message);  
-                    sendStdout(error.message);
-                    return;  
-                }  
+            // exec(cmd, { 
+            //     encoding: 'buffer',
+            //     cwd: this.execPath
+            // }, (error, stdout, stderr) => {  
+            //     if (error) {  
+            //         // logger.error(`exec error: ${iconvLite.decode(Buffer.from(error.message), 'cp936')}`);  
+            //         logger.error(error.message);  
+            //         sendStdout(error.message);
+            //         return;  
+            //     }  
 
-                if (stderr) {  
-                    logger.info(`stderr: ${iconvLite.decode(stderr, 'cp936')}`);  
-                    sendStdout(iconvLite.decode(stderr, 'cp936'));
-                }  
+            //     if (stderr) {  
+            //         logger.info(`stderr: ${iconvLite.decode(stderr, 'cp936')}`);  
+            //         sendStdout(iconvLite.decode(stderr, 'cp936'));
+            //     }  
 
-                logger.info(`stdout: ${iconvLite.decode(stdout, 'cp936')}`);  
-                sendStdout(iconvLite.decode(stdout, 'cp936'));
+            //     logger.info(`stdout: ${iconvLite.decode(stdout, 'cp936')}`);  
+            //     sendStdout(iconvLite.decode(stdout, 'cp936'));
+
+            //     if(cmd.startsWith('cd')) {
+            //         if(path) {
+            //             this.execPath = path
+            //             sendStdout('UpdateCwd:' + this.execPath);
+            //         } else {
+            //             this.execPath = resolvePath(this.execPath,(cmd.replace(/cd/,'')).trim())
+            //             sendStdout('UpdateCwd:' + this.execPath);
+            //         }
+                    
+            //     }
+               
+            // });
+            
+            // 找不到环境变量 deprecated
+            try {
+                // 假设 message 是要执行的命令  
+                const command = cmd.trim();
+
+                const child = spawn(command, command.split(" ").slice(1), {
+                    cwd: this.execPath,
+                    shell: os.platform() === 'win32'? 'powershell.exe' : true
+                });
+                // 将标准输出推送到前端  
+                child.stdout.on('data', (data) => {
+                    sendStdout(iconvLite.decode(data, 'gbk'));
+                });
+
+                // 将标准错误推送到前端（可选，但通常很有用）  
+                child.stderr.on('data', (data) => {
+                    sendStdout(iconvLite.decode(data, 'gbk'));
+                });
+
+                child.on('error', (err) => {
+                    sendStdout(err.message);
+                });
+
+                // 当命令执行完成时，可以发送一个特定的消息来通知前端  
+                child.on('close', (code) => {
+                    sendStdout(`Command finished with code ${code}`);
+                });
 
                 if(cmd.startsWith('cd')) {
                     if(path) {
@@ -89,42 +131,14 @@ class TerminalWs {
                     }
                     
                 }
-               
-            });
-
-            
-            // 找不到环境变量 deprecated
-            // try {
-            //     // 假设 message 是要执行的命令  
-            //     const command = cmd.trim();
-            //     const child = spawn(command,[],{env:{...process.env, PATH:process.env.PATH}});
-
-            //     // 将标准输出推送到前端  
-            //     child.stdout.on('data', (data) => {
-            //         sendStdout(iconvLite.decode(data, 'cp936'));
-            //     });
-
-            //     // 将标准错误推送到前端（可选，但通常很有用）  
-            //     child.stderr.on('data', (data) => {
-            //         sendStdout(iconvLite.decode(data, 'cp936'));
-            //     });
-
-            //     child.on('error', (err) => {
-            //         sendStdout(err.message);
-            //     });
-
-            //     // 当命令执行完成时，可以发送一个特定的消息来通知前端  
-            //     child.on('close', (code) => {
-            //         sendStdout(`Command finished with code ${code}`);
-            //     });
-            // } catch(err) {
-            //     if (err instanceof Error) {  
-            //         // 这里假设 zx 抛出的错误对象可能包含 stderr  
-            //         sendStdout(`Error executing command: ${err.message}\nStderr: ${(err as any).stderr || ''}`) 
-            //     } else {  
-            //         sendStdout(`Unexpected error: ${err.toString()}`);  
-            //     }  
-            // }
+            } catch(err) {
+                if (err instanceof Error) {  
+                    // 这里假设 zx 抛出的错误对象可能包含 stderr  
+                    sendStdout(`Error executing command: ${err.message}\nStderr: ${(err as any).stderr || ''}`) 
+                } else {  
+                    sendStdout(`Unexpected error: ${err.toString()}`);  
+                }  
+            }
 
         } else {
             ws.close();